Refactored examples project structure
