+++ examples/CMakeLists.txt
@@ -1,5 +1,5 @@
-add_subdirectory(c)
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/c")
 
 if(MICROSTRAIN_ENABLE_CPP)
-    add_subdirectory(cpp)
+    add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/cpp")
 endif()
