+++ examples/c/7_series/cv7/ins/cv7_ins_example.c
@@ -1,0 +1,5 @@
+// TODO: Convert C++ example to C
+int main(int argc, const char* argv[])
+{
+    return 0;
+}
