+++ examples/c/7_series/gq7/gq7_example.c
