+++ examples/c/CMakeLists.txt
@@ -1,6 +1,4 @@
-add_subdirectory(cv7)
-#add_subdirectory(cx5_gx5_45)
-#add_subdirectory(cx5_gx5_cv5_15_25)
-#add_subdirectory(gq7)
-#add_subdirectory(mip_packet)
-#add_subdirectory(stream_imu)
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/5_series")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/7_series")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/mip_packet")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/stream_imu")
