--- examples/c/cv7/CMakeLists.txt
@@ -1,46 +1,0 @@
-cmake_minimum_required(VERSION 3.16)
-
-project("Cv7_Example_C"
-    VERSION "1.0.0"
-    DESCRIPTION "Simple example for the CV7 using C"
-    LANGUAGES C
-)
-
-set(CMAKE_C_STANDARD 11)
-set(CMAKE_C_STANDARD_REQUIRED)
-
-# We need to find MicroStrain and MIP SDK in a standalone project
-if(${CMAKE_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_LIST_DIR})
-    # TODO: NOTE: Enable and update the prefix path to find the libraries in the installation directory if needed
-#    list(APPEND CMAKE_PREFIX_PATH "${CMAKE_CURRENT_BINARY_DIR}")
-
-    find_package(mip REQUIRED CONFIG)
-
-    # This example uses a serial connection and requires the microstrain_serial component
-    find_package(microstrain REQUIRED COMPONENTS microstrain_serial CONFIG)
-endif()
-
-# Make sure the MIP libraries are found
-if(NOT DEFINED CACHE{MIP_LIBRARIES} AND NOT DEFINED MIP_LIBRARIES)
-    message(FATAL_ERROR "Could not find the required MIP libraries.")
-endif()
-
-# Make sure the MicroStrain libraries are found
-if(NOT DEFINED CACHE{MICROSTRAIN_LIBRARIES} AND NOT DEFINED MICROSTRAIN_LIBRARIES)
-    message(FATAL_ERROR "Could not find the required MicroStrain libraries.")
-endif()
-
-# Create the example
-add_executable(${PROJECT_NAME}
-    "${CMAKE_CURRENT_LIST_DIR}/cv7_example.c"
-)
-
-target_link_libraries(${PROJECT_NAME} PRIVATE
-    ${MIP_LIBRARIES}
-    ${MICROSTRAIN_LIBRARIES}
-)
-
-target_include_directories(${PROJECT_NAME} PRIVATE
-    ${MIP_INCLUDE_DIRS}
-    ${MICROSTRAIN_INCLUDE_DIRS}
-)
