+++ examples/cpp/5_series/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ar_ahrs")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/gnss_ins")
