+++ examples/cpp/7_series/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/cv7")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/gq7")
