+++ examples/cpp/7_series/cv7/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ar")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ahrs")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ins")
