+++ examples/cpp/7_series/cv7/ins/cv7_ins_example.cpp
