+++ examples/cpp/7_series/gq7/gq7_example.cpp
