+++ examples/cpp/CMakeLists.txt
@@ -1,1 +1,8 @@
-add_subdirectory(cv7)
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/5_series")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/7_series")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/device_info")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/mip_packet")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/pretty_printer")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/stream_imu")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/threading")
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ublox")
