+++ examples/cpp/ublox/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory("${CMAKE_CURRENT_LIST_DIR}/cv7_ins")
